32038 - Move Magento\Framework\View\Asset\PreProcessor\Helper\SortInterface and mark as api
