+++ lib/internal/Magento/Framework/View/Asset/PreProcessor/Helper/Sort.php
@@ -5,6 +5,8 @@
  */
 namespace Magento\Framework\View\Asset\PreProcessor\Helper;
 
+use Magento\Framework\View\Asset\PreProcessor\SortInterface;
+
 /**
  * Class Sort
  */
